% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{print.mak}
\alias{print.mak}
\title{Print 'mak' object}
\usage{
\method{print}{mak}(x, ...)
}
\arguments{
\item{x}{Object of class \code{mak}.}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\description{
Prints objects of class \code{mak}, including visualising the reactions.
}
\examples{
# Michaelis-Menten system
A <- matrix(
c(1, 1, 0, 0,
  0, 0, 1, 0,
  0, 0, 1, 0), ncol = 4, byrow = TRUE)
B <- matrix(
c(0, 0, 1, 0,
  1, 1, 0, 0,
  1, 0, 0, 1), ncol = 4, byrow = TRUE)
m <- mak(A, B)
m
}
\seealso{
mak
}
