% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{print.ode}
\alias{print.ode}
\title{Print 'ode' object}
\usage{
\method{print}{ode}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ode}.}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\description{
Prints objects of class \code{ode}. Since this class is abstract, it moves on the print method for the derived class.
}
\seealso{
ode
}
