% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{print.plk}
\alias{print.plk}
\title{Print 'plk' object}
\usage{
\method{print}{plk}(x, ...)
}
\arguments{
\item{x}{Object of class \code{plk}.}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\description{
Prints objects of class \code{plk}, including visualising the powers.
}
\examples{
# Power law system
A <- matrix(
c(1, 0, 0, 0,
  0, 1, 2, 0,
  0, 0, 0, 0), ncol = 4, byrow = TRUE)
p <- plk(A)
p
}
\seealso{
plk
}
