% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.R
\name{print.reg}
\alias{print.reg}
\title{Print 'reg' object}
\usage{
\method{print}{reg}(x, ...)
}
\arguments{
\item{x}{Object of class \code{reg}.}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\description{
Prints objects of class \code{reg}.
}
\examples{
# Example: default regularisation
reg()

}
\seealso{
reg
}
