% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{print.rlk}
\alias{print.rlk}
\title{Print 'rlk' object}
\usage{
\method{print}{rlk}(x, ...)
}
\arguments{
\item{x}{Object of class \code{rlk}.}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\description{
Prints objects of class \code{rlk} and presents the fractions.
}
\examples{
# Rational law kinetics
A <- matrix(
c(1, 0, 0, 0,
  0, 1, 2, 0,
  0, 0, 0, 1), ncol = 4, byrow = TRUE)
r <- rlk(A, A[c(2, 1, 3), ])
r
}
\seealso{
rlk
}
