% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{print.solver}
\alias{print.solver}
\title{Print 'solver' object}
\usage{
\method{print}{solver}(x, ...)
}
\arguments{
\item{x}{Object of class \code{solver}.}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\description{
Prints objects of class \code{solver}.
}
\examples{
# Dormand-Prince 4/5 scheme
solver("dp45")

# Runge-Kutta-Fehlberg 4/5 scheme (default)
solver()

}
\seealso{
solver
}
